chore: move search function to searchUtil as it is shared func
